Separated external API calls
